dcp-712 Move util function to entity map test suite
