+++ tests/unit/importer/submission/test_submission.py
@@ -71,74 +71,3 @@ class SubmissionTest(TestCase):
         self.assertEqual(count_per_entity['protocol'], submitted_json['expectedProtocols'])
         self.assertEqual(count_per_entity['project'], submitted_json['expectedProjects'])
         self.assertEqual(total_count, submitted_json['expectedLinks'])
-
-
-def _create_spreadsheet_json():
-    spreadsheet_json = {
-        'project': {
-            'dummy-project-id': {
-                'content': {
-                    'key': 'project_1'
-                }
-            }
-        },
-        'biomaterial': {
-            'biomaterial_id_1': {
-                'content': {
-                    'key': 'biomaterial_1'
-                }
-            },
-            'biomaterial_id_2': {
-                'content': {
-                    'key': 'biomaterial_2'
-                },
-                'links_by_entity': {
-                    'biomaterial': ['biomaterial_id_1'],
-                    'process': ['process_id_1']
-                }
-            },
-            'biomaterial_id_3': {
-                'content': {
-                    'key': 'biomaterial_3'
-                },
-                'links_by_entity': {
-                    'biomaterial': ['biomaterial_id_2'],
-                    'process': ['process_id_2']
-                }
-            },
-            'biomaterial_id_4': {
-                'content': {
-                    'key': 'biomaterial_3'
-                },
-                'links_by_entity': {
-                    'biomaterial': ['biomaterial_id_2'],
-                    'process': ['process_id_2']
-                },
-                'external_links_by_entity': {
-                    'biomaterial': ['biomaterial_uuid']
-                },
-
-            },
-        },
-        'file': {
-            'file_id_1': {
-                'content': {
-                    'file_core': {
-                        'file_name': 'file_name'
-                    }
-                },
-                'links_by_entity': {
-                    'biomaterial': ['biomaterial_id_3']
-                }
-            }
-        },
-        'protocol': {
-            'protocol_id_1': {
-                'content': {
-                    'key': 'protocol_1'
-                }
-            }
-        }
-    }
-
-    return spreadsheet_json
